fix: age 자료형 변경

+++ src/main/kotlin/sonny/kotlinspringprac/web/res/UserResponse.kt
@@ -5,7 +5,7 @@ import sonny.kotlinspringprac.domain.Post
 data class UserResponse(
     val id: String,
     val name: String,
-    val age: Long,
+    val age: Int,
     val email: String,
     val posts: List<Post>? = null,
 )
